Build tweaks
* Removed build from tox.ini; handled in the Makefile now.

* Added a few files into charmcraft.yaml to be included during the
  build process.  In turn, removed manual copying of those files into
  the unpacked charm via the "unpack" Makefile step.

Note that the above makes the "unpack" step purely present for
powering functional tests.  It likely needs to be fixed anyway due to
charm naming in newer versions of charmcraft, but if we can confirm
that things work with the .charm files instead, we can remove the
unpack step completely.